+++ plugins/doc_fragments/nios.py
@@ -81,7 +81,7 @@ options:
             variable.
           - Until ansible 2.8 the default WAPI was 1.4
         type: str
-        default: '2.1'
+        default: '2.9'
       max_results:
         description:
           - Specifies the maximum number of objects to be returned,